Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -43,7 +43,7 @@ int main()
     }
     else if (c == '*')
     {
-        printf("%d", (float)(a * b));  // Multiplication (output as float)
+        printf("%d", (a * b));  // Multiplication (output as float)
     }
     else if (c == '/')
     {
